// autogpt_platform/frontend/src/app/api/__generated__/models/mCPToolOutputResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { MCPToolOutputResponseSessionId } from "./mCPToolOutputResponseSessionId";

/**
 * Response after executing an MCP tool.
 */
export interface MCPToolOutputResponse {
  type?: ResponseType;
  message: string;
  session_id?: MCPToolOutputResponseSessionId;
  server_url: string;
  tool_name: string;
  result?: unknown;
  success?: boolean;
}